+++ packages/shorebird_cli/bin/shorebird.dart
@@ -23,7 +23,6 @@ import 'package:shorebird_cli/src/shorebird_android_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_flutter.dart';
-import 'package:shorebird_cli/src/shorebird_logger.dart';
 import 'package:shorebird_cli/src/shorebird_process.dart';
 import 'package:shorebird_cli/src/shorebird_validator.dart';
 import 'package:shorebird_cli/src/shorebird_version.dart';
@@ -63,7 +62,6 @@ Future<void> main(List<String> args) async {
         shorebirdArtifactsRef,
         shorebirdEnvRef,
         shorebirdFlutterRef,
-        shorebirdLoggerRef,
         shorebirdValidatorRef,
         shorebirdVersionRef,
         xcodeBuildRef,
