+++ packages/shorebird_cli/test/src/command_runner_test.dart
@@ -8,7 +8,6 @@ import 'package:shorebird_cli/src/logger.dart' hide logger;
 import 'package:shorebird_cli/src/platform.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_flutter.dart';
-import 'package:shorebird_cli/src/shorebird_logger.dart';
 import 'package:shorebird_cli/src/shorebird_version.dart';
 import 'package:shorebird_cli/src/version.dart';
 import 'package:test/test.dart';
@@ -21,10 +20,9 @@ void main() {
     const flutterRevision = 'test-flutter-revision';
     const flutterVersion = '1.2.3';
 
-    late Logger logger;
+    late ShorebirdLogger logger;
     late Platform platform;
     late ShorebirdEnv shorebirdEnv;
-    late ShorebirdLogger shorebirdLogger;
     late ShorebirdFlutter shorebirdFlutter;
     late ShorebirdVersion shorebirdVersion;
     late ShorebirdCliCommandRunner commandRunner;
@@ -38,18 +36,21 @@ void main() {
           shorebirdEnvRef.overrideWith(() => shorebirdEnv),
           shorebirdFlutterRef.overrideWith(() => shorebirdFlutter),
           shorebirdVersionRef.overrideWith(() => shorebirdVersion),
-          shorebirdLoggerRef.overrideWith(() => shorebirdLogger),
         },
       );
     }
 
     setUp(() {
-      logger = MockLogger();
+      logger = MockShorebirdLogger();
       platform = MockPlatform();
       shorebirdEnv = MockShorebirdEnv();
       shorebirdFlutter = MockShorebirdFlutter();
       shorebirdVersion = MockShorebirdVersion();
       when(() => logger.level).thenReturn(Level.info);
+      final logFile = MockFile();
+      when(() => logger.logFile).thenReturn(logFile);
+      when(() => logFile.absolute).thenReturn(logFile);
+      when(() => logFile.path).thenReturn('test.log');
       when(
         () => shorebirdEnv.shorebirdEngineRevision,
       ).thenReturn(shorebirdEngineRevision);
@@ -60,11 +61,6 @@ void main() {
       ).thenAnswer((_) async => flutterVersion);
       when(() => shorebirdVersion.isLatest()).thenAnswer((_) async => true);
       commandRunner = runWithOverrides(ShorebirdCliCommandRunner.new);
-      shorebirdLogger = MockShorebirdLogger();
-      final logFile = MockFile();
-      when(() => shorebirdLogger.logFile).thenReturn(logFile);
-      when(() => logFile.absolute).thenReturn(logFile);
-      when(() => logFile.path).thenReturn('test.log');
     });
 
     test('handles FormatException', () async {
@@ -289,8 +285,7 @@ Run ${lightCyan.wrap('shorebird upgrade')} to upgrade.'''),
         await runWithOverrides(
           () => commandRunner.run(['release', 'android', '--verbose']),
         );
-        verify(() => shorebirdLogger.detail(any(that: contains('#0'))))
-            .called(1);
+        verify(() => logger.detail(any(that: contains('#0')))).called(1);
       });
 
       group('when running with --verbose', () {
