+++ packages/shorebird_cli/test/src/commands/patch/patch_aar_command_test.dart
@@ -98,7 +98,7 @@ void main() {
     late PatchDiffChecker patchDiffChecker;
     late Platform platform;
     late Progress progress;
-    late Logger logger;
+    late ShorebirdLogger logger;
     late ShorebirdProcessResult flutterBuildProcessResult;
     late ShorebirdProcessResult flutterPubGetProcessResult;
     late http.Client httpClient;
@@ -186,7 +186,7 @@ void main() {
       patchDiffChecker = MockPatchDiffChecker();
       platform = MockPlatform();
       progress = MockProgress();
-      logger = MockLogger();
+      logger = MockShorebirdLogger();
       flutterBuildProcessResult = MockProcessResult();
       flutterPubGetProcessResult = MockProcessResult();
       httpClient = MockHttpClient();
