+++ packages/shorebird_cli/test/src/commands/release_new/android_releaser_test.dart
@@ -36,7 +36,7 @@ void main() {
     late Doctor doctor;
     late Platform platform;
     late Directory projectRoot;
-    late Logger logger;
+    late ShorebirdLogger logger;
     late OperatingSystemInterface operatingSystemInterface;
     late Progress progress;
     late ShorebirdFlutterValidator flutterValidator;
@@ -85,7 +85,7 @@ void main() {
       platform = MockPlatform();
       progress = MockProgress();
       projectRoot = Directory.systemTemp.createTempSync();
-      logger = MockLogger();
+      logger = MockShorebirdLogger();
       flutterValidator = MockShorebirdFlutterValidator();
       shorebirdProcess = MockShorebirdProcess();
       shorebirdEnv = MockShorebirdEnv();
