+++ packages/shorebird_cli/test/src/commands/release_new/ios_framework_releaser_test.dart
@@ -38,7 +38,7 @@ void main() {
       late Doctor doctor;
       late Platform platform;
       late Directory projectRoot;
-      late Logger logger;
+      late ShorebirdLogger logger;
       late OperatingSystemInterface operatingSystemInterface;
       late Progress progress;
       late ShorebirdFlutterValidator flutterValidator;
@@ -87,7 +87,7 @@ void main() {
         platform = MockPlatform();
         progress = MockProgress();
         projectRoot = Directory.systemTemp.createTempSync();
-        logger = MockLogger();
+        logger = MockShorebirdLogger();
         shorebirdProcess = MockShorebirdProcess();
         shorebirdEnv = MockShorebirdEnv();
         flutterValidator = MockShorebirdFlutterValidator();
